+++ crates/alkahest-renderer/src/renderer/mod.rs
@@ -534,6 +534,7 @@ pub enum RenderDebugView {
     SpecularLight,
     SpecularOcclusion,
     SpecularSmoothness,
+    SpecularOnly,
 
     Emissive,
     EmissiveIntensity,
@@ -545,6 +546,11 @@ pub enum RenderDebugView {
     DepthEdges,
     DepthGradient,
     DepthWalkable,
+
+    ValidLayeredMetalness,
+    ValidSmoothnessHeatmap,
+    ValidSourceColorBrightness,
+    ValidSourceColorSaturation,
 }
 
 impl RenderDebugView {